Fix constructor calls for ChaveSeletora and Botao classes
This commit fixes the errors that occurred when calling the constructors for the ChaveSeletora and Botao classes. The constructors were being called with an incorrect number of arguments. The ChaveSeletora class constructor now receives 3 arguments: an integer representing the pin number of the switch, an unsigned long representing the debounce delay in milliseconds, and a pointer to an LED object that the switch controls. The Botao class constructor now receives 2 arguments: an integer representing the pin number of the button and an unsigned long representing the debounce delay in milliseconds.